% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDailyStationData.R
\name{dtGetDailyStationDataOverUpToOneYear}
\alias{dtGetDailyStationDataOverUpToOneYear}
\title{Fetch daily Meteostat data for a station}
\usage{
dtGetDailyStationDataOverUpToOneYear(
  iMeteostatStationId,
  cApiKey,
  dateStartDate,
  dateEndDate
)
}
\arguments{
\item{iMeteostatStationId}{- integer scalar, ID of the station as given by
Meteostat's dictionary table of weather stations}

\item{cApiKey}{- character scalar, Meteostat's key to use to fetch the data}

\item{dateStartDate}{- Date class scalar, first date of the interval for which
the data is to be obtained}

\item{dateEndDate}{- Date class scalar, last date of the interval for which
the data is to be obtained}
}
\value{
data.table with the data fetched; the colnames are self-explanatory,
in case of doubt cf. the implementation and API's documentation:
https://dev.meteostat.net/api/stations/daily.html#response
}
\description{
Fetch daily Meteostat data for a station for at most yearly interval
}
\details{
This function fetches data made available via Meteostat's API. Please note that
due to the cap on the size of the window for which the data can be fetched,
this function can only handle requests that ask for data for up to 370 days.
For generalized version of this function capable of handling arbitrary
time intervals cf. dtGetDailyStationData
}
