% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationsFunctions.R
\name{dtGetMeteostatWeatherStationsDict}
\alias{dtGetMeteostatWeatherStationsDict}
\title{Download the full list set of stations}
\usage{
dtGetMeteostatWeatherStationsDict()
}
\value{
data.table with the following columns:
}
\description{
Download the full list of Meteostat's stations (updated daily)
}
\details{
This function downloads and parses the file with the full list of meteo stations
that feed data to Meteostat's API. Cf.:
https://dev.meteostat.net/bulk/stations.html#endpoints
}
